skip extracting if exist

+++ itraq/entry.py
@@ -32,12 +32,13 @@ for src_name in glob.glob(os.path.join(working_dir, '*.gz')):
     base = os.path.basename(src_name)
     print("Extracting", src_name)
     dest_name = os.path.join(working_dir, base[:-3])
-    spectrum_tmp.append(dest_name)
-    with gzip.open(src_name, 'rb') as infile:
-        with open(dest_name, 'wb') as outfile:
-            for line in infile:
-                outfile.write(line)
-
+    if not os.path.isfile(dest_name):
+    	spectrum_tmp.append(dest_name)
+    	with gzip.open(src_name, 'rb') as infile:
+	        with open(dest_name, 'wb') as outfile:
+	            for line in infile:
+	                outfile.write(line)
+	                
 spectrum 				= scan_spectrum(working_dir)
 db, input_csv 			= scan_dir(working_dir)
 #module 		= dir_content['module']
