+++ scquant/entry.py
@@ -2,6 +2,7 @@
 # EXTRACT, IF NECESSARY, AND RUN MGSF PLUS TO 
 # NORMALIZE THE DATA
 # DANIEL.KRISTIYANTO@PNNL.GOV
+# Docker Container: scquant
 
 ######################## ENV ###########################
 import gzip
@@ -13,6 +14,8 @@ import re
 import csv
 from ftplib import FTP
 from _functions import *
+from datetime import timedelta
+import time
 
 working_dir = os.getcwd() +"/data/"
 
@@ -28,11 +31,12 @@ for src_name in glob.glob(os.path.join(working_dir, '*.gz')):
     base = os.path.basename(src_name)
     print("Extracting", src_name)
     dest_name = os.path.join(working_dir, base[:-3])
-    spectrum_tmp.append(dest_name)
-    with gzip.open(src_name, 'rb') as infile:
-        with open(dest_name, 'wb') as outfile:
-            for line in infile:
-                outfile.write(line)
+    if not os.path.isfile(dest_name):
+    	spectrum_tmp.append(dest_name)
+    	with gzip.open(src_name, 'rb') as infile:
+	        with open(dest_name, 'wb') as outfile:
+	            for line in infile:
+	                outfile.write(line)
 
 spectrum 				= scan_spectrum(working_dir)
 db, input_csv 			= scan_dir(working_dir)
